+++ src/lib/pages/pools/components/pool-details/tables/pool-txs/PoolTxsMsg.tsx
@@ -36,7 +36,7 @@ export const PoolTxsMsg = ({
   const { isOpen, onToggle } = useDisclosure();
   const msgDetailTemplateColumns = templateColumns
     .split(" ")
-    .slice(0, 2)
+    .slice(0, 3)
     .concat("1fr")
     .join(" ");
   return (
@@ -146,6 +146,7 @@ export const PoolTxsMsg = ({
           hidden={!isOpen}
           templateColumns={msgDetailTemplateColumns}
         >
+          <TableNoBorderRow py={0} />
           <TableNoBorderRow py={0} />
           <TableNoBorderRow py={0} />
           <TableNoBorderRow py={0}>