+++ src/lib/pages/pools/components/pool-details/tables/pool-txs/messages/swap/MsgSwapExactAmountOutDetail.tsx
@@ -1,7 +1,9 @@
-import { Box, Flex, Text } from "@chakra-ui/react";
+import { Box, Flex } from "@chakra-ui/react";
 
+import { PoolInfoText } from "../components/PoolInfoText";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import type { AssetInfosOpt } from "lib/services/assetService";
+import { extractMsgType } from "lib/utils";
 import type { MsgSwapExactAmountOutDetails } from "lib/utils/tx/types";
 
 import { PoolRoute, PoolSwap } from "./components";
@@ -34,17 +36,17 @@ export const MsgSwapExactAmountOutDetail = ({
   }));
 
   return (
-    <Flex w="full" alignItems="start" gap={12}>
-      <Flex direction="column" gap={1}>
-        <Text variant="body2" textColor="pebble.500" fontWeight={500}>
-          Block height
-        </Text>
-        <ExplorerLink
-          value={blockHeight.toString()}
-          type="block_height"
-          showCopyOnHover
-          ampCopierSection={ampCopierSection}
-        />
+    <Flex w="full" direction="column" alignItems="start" gap={6}>
+      <Flex gap={12}>
+        <PoolInfoText title="Block height">
+          <ExplorerLink
+            value={blockHeight.toString()}
+            type="block_height"
+            showCopyOnHover
+            ampCopierSection={ampCopierSection}
+          />
+        </PoolInfoText>
+        <PoolInfoText title="Message">{extractMsgType(msg.type)}</PoolInfoText>
       </Flex>
       <Box w="full">
         <PoolSwap